pyproject.toml
[project]
name = "uneseule-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "fastapi>=0.124.4",
    "httpx>=0.28.1",
    "motor>=3.7.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "strawberry-graphql[fastapi]>=0.254.0",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.9",
]
